create components package in main app, refactor existing code into it
