+++ app/src/main/java/org/torproject/android/ui/more/camo/CamoFragment.kt
@@ -15,7 +15,7 @@ import org.torproject.android.R
 import org.torproject.android.core.getKey
 import org.torproject.android.service.util.Prefs
 import org.torproject.android.ui.more.MoreActionAdapter
-import org.torproject.android.ui.OrbotMenuAction
+import org.torproject.android.ui.components.OrbotMenuAction
 import kotlin.String
 
 class CamoFragment : Fragment() {
